requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
pycparser==2.21
Werkzeug==3.0.1
